# Shallow copy where possible in data processing (15% faster)